Reuse previous connection
The gui.py file now looks for a json file in the working directory which contains the previous spreadsheet_id, trello key and trello token. It then uses those informations to fill in automatically the entry boxes. The user may use that same info or change it to new ones. Whatever info is used is then written on top of the old json file.